+++ attbacktrader/data/__init__.py
@@ -7,11 +7,18 @@ from .industries import ShenwanIndustryClassification, StockIndustryMembership
 from .quality import DataQualityIssue, assess_daily_bar_quality
 from .resampling import resample_daily_bars
 from .stock_pool import (
+    DynamicStockPool,
+    DynamicStockPoolEntry,
+    DynamicStockPoolMembership,
     FixedStockPoolMember,
     IndexConstituent,
+    dynamic_stock_pool_entries_from_index_constituents,
     fixed_stock_pool_members_from_index_constituents,
+    fixed_stock_pool_members_from_dynamic_entries,
     latest_index_constituents,
+    read_dynamic_stock_pool_parquet,
     read_fixed_stock_pool_csv,
+    write_dynamic_stock_pool_parquet,
     write_fixed_stock_pool_csv,
 )
 from .tradability import TradabilityStatus
@@ -19,6 +26,9 @@ from .tradability import TradabilityStatus
 __all__ = [
     "DataQualityIssue",
     "DailyBar",
+    "DynamicStockPool",
+    "DynamicStockPoolEntry",
+    "DynamicStockPoolMembership",
     "FixedStockPoolMember",
     "IndexBar",
     "IndexConstituent",
@@ -28,10 +38,14 @@ __all__ = [
     "TradingCalendar",
     "TradabilityStatus",
     "assess_daily_bar_quality",
+    "dynamic_stock_pool_entries_from_index_constituents",
     "fixed_stock_pool_members_from_index_constituents",
+    "fixed_stock_pool_members_from_dynamic_entries",
     "latest_index_constituents",
+    "read_dynamic_stock_pool_parquet",
     "read_fixed_stock_pool_csv",
     "resample_daily_bars",
     "trading_calendar_from_bars",
+    "write_dynamic_stock_pool_parquet",
     "write_fixed_stock_pool_csv",
 ]